Add new class Service Use snake_case for names of files
